+++ cmd/tasks/rds/tags_test.go
@@ -1,4 +1,4 @@
-package main
+package rds
 
 import (
 	"testing"
