+++ cmd/tasks/tags/tags.go
@@ -1,4 +1,4 @@
-package main
+package tags
 
 import (
 	"log"
